Share account API across node and web SDKs
